pyproject.toml
[project]
name = "personal-agent"
version = "0.1.0"
description = "个人知识库智能助手 - RAG + 长期记忆 + 个性化对话"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-community>=0.0.10",
    "pymilvus>=2.3.0",
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.5.0",
    "pyyaml>=6.0.1",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]